update
E/AndroidRuntime: FATAL EXCEPTION: main
    Process: haina.ecommerce, PID: 25657
    java.lang.IllegalStateException: Fragment FillDataPassengerFragment{94783fd} (9dedc564-a872-49c8-a4ed-fd59cb2ebbe2)} not attached to an activity.
        at androidx.fragment.app.Fragment.requireActivity(Fragment.java:833)
        at haina.ecommerce.view.flight.filldatapassenger.FillDataPassengerFragment.getCalculationPrice(FillDataPassengerFragment.kt:231)
        at haina.ecommerce.view.flight.filldatapassenger.FillDataPassengerPresenter$getCalculationTicketPrice$1.onResponse(FillDataPassengerPresenter.kt:19)
        at retrofit2.DefaultCallAdapterFactory$ExecutorCallbackCall$1$1.run(DefaultCallAdapterFactory.java:83)
        at android.os.Handler.handleCallback(Handler.java:938)
        at android.os.Handler.dispatchMessage(Handler.java:99)
        at android.os.Looper.loop(Looper.java:223)
        at android.app.ActivityThread.main(ActivityThread.java:7656)
        at java.lang.reflect.Method.invoke(Native Method)
        at com.android.internal.os.RuntimeInit$MethodAndArgsCaller.run(RuntimeInit.java:592)
        at com.android.internal.os.ZygoteInit.main(ZygoteInit.java:947)

+++ app/src/main/java/haina/ecommerce/view/flight/filldatapassenger/FillDataPassengerFragment.kt
@@ -172,7 +172,9 @@ class FillDataPassengerFragment : Fragment(), View.OnClickListener,
 
     override fun onDestroy() {
         super.onDestroy()
-        deleteAllPassenger()
+        if (listDataPassenger.isNotEmpty()){
+            deleteAllPassenger()
+        }
     }
 
     private fun setlistTicket(){
